Update core database setup to support environment-specific seeding
