+++ script/bootstrap
@@ -24,4 +24,5 @@ npm install
 echo "==> Creating and setting up databases..."
 npm run db:create
 
+echo ""
 echo '‼️ Create a ".env" file based on the ".env.example" file, then enter all of your environment variables.'
