+++ script/create-db.js
@@ -1,3 +1,5 @@
+const fs = require('fs')
+const path = require('path')
 const { createClient, getConfig, getEnv } = require('../server/db')
 
 // See https://www.postgresql.org/docs/13/errcodes-appendix.html
@@ -14,7 +16,7 @@ async function bootstrapDatabase() {
   const config = getConfig(targetEnv)
   await createDatabase(config)
   await migrateToLatestSchemas(config)
-  //await runDataSeeders(config)
+  await runDataSeeders(config)
 }
 
 async function createDatabase(config) {
@@ -71,7 +73,27 @@ async function runDataSeeders(config) {
   const { database } = config.connection
   let db
 
+  if (!config || !config.seeds || !config.seeds.directory) {
+    console.warn('Skipping! No data seed directory is configured.')
+    return
+  }
+
+  const seedDir = path.resolve(process.cwd(), config.seeds.directory)
+  const seedDirStats = fs.statSync(seedDir, { throwIfNoEntry: false })
+  if (!seedDirStats) {
+    console.warn(
+      `Skipping! The data seed directory does not exist: ${config.seeds.directory}`
+    )
+    return
+  }
+
   try {
+    if (!seedDirStats.isDirectory()) {
+      throw new Error(
+        `The data seed path exists but is not a directory: ${config.seeds.directory}`
+      )
+    }
+
     db = createClient(config)
     await db.seed.run()
     console.log(`Created seed data in database "${database}"!`)
@@ -82,6 +104,8 @@ async function runDataSeeders(config) {
     throw error
   } finally {
     // Disconnect
-    await db.destroy()
+    if (db) {
+      await db.destroy()
+    }
   }
 }
