+++ server/db/seeds/development/seed-initial-data.js
@@ -1,6 +1,7 @@
 module.exports = {
   async seed(knex) {
     // Deletes ALL existing entries
+    await knex('letter_versions').del() // Because these have foreign keys linked to the campaigns table
     await knex('campaigns').del()
 
     // Inserts seed entries
